# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FOODEX API
  version: 0.2.0
  description: Implementation baseline for FOODEX v1 mobile and web clients.
  license:
    name: Proprietary
servers:
  - url: /api/v1
security: []
...
paths:
...
  /quote:
    post:
      tags: [Checkout]
      summary: Reprice the authenticated store cart using backend-authoritative commercial rules
      operationId: quoteCart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequest"
      responses:
        "200":
          description: Authoritative quote. Clients must treat all returned amounts as read-only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /checkout:
    post:
      tags: [Checkout]
      summary: Create an order from the authenticated cart
      operationId: checkout
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Client-generated key preventing duplicate order creation for a retried checkout.
      schema:
        type: string
        minLength: 16
        maxLength: 100
  schemas:
...
    QuoteRequest:
      type: object
      additionalProperties: false
      required: [store_id]
      properties:
        store_id:
          type: integer
          minimum: 1
        coupon_code:
          type: [string, "null"]
          maxLength: 80
        payment_method:
          type: [string, "null"]
          maxLength: 50
    QuoteLine:
      type: object
      required:
        [product_id, quantity, unit_price, line_total, is_available]
      properties:
        product_id:
          type: integer
        sku:
          type: [string, "null"]
        name:
          type: [string, "null"]
        quantity:
          type: number
          exclusiveMinimum: 0
          multipleOf: 0.001
        base_unit_price:
          type: [number, "null"]
          multipleOf: 0.001
        unit_price:
          type: [number, "null"]
          multipleOf: 0.001
        line_total:
          type: [number, "null"]
          multipleOf: 0.001
        is_available:
          type: boolean
        available_quantity:
          type: [number, "null"]
          minimum: 0
        minimum_order_quantity:
          type: [number, "null"]
        ordering_increment:
          type: [number, "null"]
        pack_size:
          type: [number, "null"]
        case_size:
          type: [number, "null"]
        price_tier_id:
          type: [integer, "null"]
        price_tier_code:
          type: [string, "null"]
    CommerceQuote:
      type: object
      required:
        [quote_id, quoted_at, store_id, channel, currency, items, subtotal, discount_total, delivery_total, tax_total, grand_total, pricing_source]
      properties:
        quote_id:
          type: string
          format: uuid
        quoted_at:
          type: string
          format: date-time
        store_id:
          type: integer
        channel:
          type: string
          enum: [b2b, b2c]
        currency:
          type: string
          enum: [EGP]
        customer_context:
          type: object
          properties:
            customer_id:
              type: [integer, "null"]
            b2b_account_id:
              type: [integer, "null"]
            price_tier_id:
              type: [integer, "null"]
            price_tier_code:
              type: [string, "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLine"
        subtotal:
          type: number
          multipleOf: 0.001
        promotion_discount_total:
          type: number
          multipleOf: 0.001
        coupon_discount_total:
          type: number
          multipleOf: 0.001
        discount_total:
          type: number
          multipleOf: 0.001
        delivery_total:
          type: number
          multipleOf: 0.001
        tax_rate:
          type: number
          minimum: 0
        tax_total:
          type: number
          multipleOf: 0.001
        grand_total:
          type: number
          multipleOf: 0.001
        has_unavailable_items:
          type: boolean
        pricing_source:
          type: string
          enum: [backend_authoritative_quote_v1]
    QuoteResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CommerceQuote"
    CheckoutRequest:
      type: object
      additionalProperties: false
      required: [store_id, address_id]
      properties:
        store_id:
          type: integer
          minimum: 1
        address_id:
          type: integer
          minimum: 1
        payment_method:
          type: [string, "null"]
        note:
          type: [string, "null"]
          maxLength: 1000
    OrderItem:
      type: object
      required: [id, product_id, sku, name, quantity, unit_price, line_total]
      properties:
        id:
          type: integer
        product_id:
          type: integer
        sku:
          type: string
        name:
          type: string
        quantity:
          type: number
          exclusiveMinimum: 0
          multipleOf: 0.001
        unit_price:
          type: number
          multipleOf: 0.001
        line_total:
          type: number
          multipleOf: 0.001
    Order:
      type: object
      required:
        [id, order_number, store_id, store, channel, status, currency, subtotal, grand_total, items, created_at]
      properties:
        id:
          type: integer
        order_number:
          type: string
        store_id:
          type: integer
        store:
          type: [object, "null"]
          properties:
            id:
              type: integer
            code:
              type: string
            name:
              type: string
            logo_url:
              type: [string, "null"]
        channel:
          type: string
          enum: [b2c, b2b]
        status:
          type: string
        currency:
          type: string
          enum: [EGP]
        subtotal:
          type: number
          multipleOf: 0.001
        discount_total:
          type: number
          multipleOf: 0.001
        delivery_total:
          type: number
          multipleOf: 0.001
        tax_total:
          type: number
          multipleOf: 0.001
        quote_id:
          type: [string, "null"]
          format: uuid
        quoted_at:
          type: [string, "null"]
          format: date-time
        price_tier_id:
          type: [integer, "null"]
        price_tier_code:
          type: [string, "null"]
        grand_total:
          type: number
          multipleOf: 0.001
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        invoice:
          type: [object, "null"]
          properties:
            id:
              type: integer
            invoice_number:
              type: string
            status:
              type: string
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: [string, "null"]
        errors:
          type: [object, "null"]
          additionalProperties: true
  responses:
...
    Unauthenticated:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found or not visible in authorized scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state or idempotency key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...